correcoes telas jogos, fragments
