--- app/src/main/java/br/com/digitalhouse/harvardartmuseums/fragments/game/GameControlsActivity.java
@@ -1,60 +1,0 @@
-package br.com.digitalhouse.harvardartmuseums.fragments.game;
-
-import android.content.Intent;
-import androidx.appcompat.app.AppCompatActivity;
-import android.os.Bundle;
-import androidx.appcompat.widget.Toolbar;
-import android.view.Menu;
-import android.view.MenuItem;
-
-import br.com.digitalhouse.harvardartmuseums.R;
-import br.com.digitalhouse.harvardartmuseums.view.help.HelpActivity;
-import br.com.digitalhouse.harvardartmuseums.view.settings.SettingsActivity;
-
-public class GameControlsActivity extends AppCompatActivity {
-
-    private Toolbar toolbar;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_game_controls);
-
-        toolbar = findViewById(R.id.toolbar);
-        toolbar.setTitle("Art memories - Controls");
-        setSupportActionBar(toolbar);
-
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_main, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
-
-        //noinspection SimplifiableIfStatement
-        if (id == R.id.action_help) {
-            Intent intent = new Intent(GameControlsActivity.this, HelpActivity.class);
-            startActivity(intent);
-        }
-
-        if (id == R.id.action_settings) {
-            Intent intent = new Intent(GameControlsActivity.this, SettingsActivity.class);
-            startActivity(intent);
-        }
-
-        return super.onOptionsItemSelected(item);
-    }
-
-
-}
-
-
