--- app/src/main/java/br/com/digitalhouse/harvardartmuseums/fragments/game/GamePlay1Activity.java
@@ -1,70 +1,0 @@
-package br.com.digitalhouse.harvardartmuseums.fragments.game;
-
-import android.content.Intent;
-import androidx.appcompat.app.AppCompatActivity;
-import android.os.Bundle;
-import androidx.appcompat.widget.Toolbar;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.ImageView;
-
-import br.com.digitalhouse.harvardartmuseums.R;
-import br.com.digitalhouse.harvardartmuseums.view.help.HelpActivity;
-import br.com.digitalhouse.harvardartmuseums.view.settings.SettingsActivity;
-
-public class GamePlay1Activity extends AppCompatActivity {
-
-    private Toolbar toolbar;
-    private ImageView imageViewGame;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_game_play1);
-
-        toolbar = findViewById(R.id.toolbar);
-        toolbar.setTitle("Art memories");
-        setSupportActionBar(toolbar);
-
-        //Inicialização da view ref. ao processamento do jogo (game) e Inserção de ação para mudar para a activity do resultado do jogo
-        imageViewGame = findViewById(R.id.imageViewGame);
-
-        imageViewGame.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(GamePlay1Activity.this,GamePlay2Activity.class);
-                startActivity(intent);
-            }
-        });
-
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_main, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
-
-        //noinspection SimplifiableIfStatement
-        if (id == R.id.action_help) {
-            Intent intent = new Intent(GamePlay1Activity.this, HelpActivity.class);
-            startActivity(intent);
-        }
-
-        if (id == R.id.action_settings) {
-            Intent intent = new Intent(GamePlay1Activity.this, SettingsActivity.class);
-            startActivity(intent);
-        }
-
-        return super.onOptionsItemSelected(item);
-    }
-}
