+++ app/src/main/java/br/com/digitalhouse/harvardartmuseums/fragments/home/HomeFragment.java
@@ -2,11 +2,14 @@ package br.com.digitalhouse.harvardartmuseums.fragments.home;
 
 import android.content.Context;
 import android.os.Bundle;
+
 import androidx.fragment.app.Fragment;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
+import android.widget.TextView;
 
 import br.com.digitalhouse.harvardartmuseums.R;
 import br.com.digitalhouse.harvardartmuseums.interfaces.Comunicator;
@@ -14,6 +17,9 @@ import br.com.digitalhouse.harvardartmuseums.interfaces.Comunicator;
 public class HomeFragment extends Fragment {
 
     private Comunicator comunicator;
+
+    private TextView textViewTopMessageChoose;
+    private ImageView imageViewTopMessage;
 
     private ImageView subsolo;
     private ImageView andar1;
@@ -32,16 +38,21 @@ public class HomeFragment extends Fragment {
 
         try {
             comunicator = (Comunicator) context;
-        }catch (Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
         }
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+
         View view = inflater.inflate(R.layout.fragment_home, container, false);
 
+        //Inicializa Views
         initViews(view);
+
+        textViewTopMessageChoose.setText("Choose a floor to visit the Gallery");
+        imageViewTopMessage.setImageDrawable(getResources().getDrawable(R.drawable.personagem04));
 
         subsolo.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -84,12 +95,17 @@ public class HomeFragment extends Fragment {
     }
 
     private void initViews(View view) {
+
+        textViewTopMessageChoose = view.findViewById(R.id.textViewTopMessageChoose);
+        imageViewTopMessage = view.findViewById(R.id.imageViewTopMessage);
+
         subsolo = view.findViewById(R.id.imageViewAndarSS);
         andar1 = view.findViewById(R.id.imageViewAndar1);
         andar2 = view.findViewById(R.id.imageViewAndar2);
         andar3 = view.findViewById(R.id.imageViewAndar3);
         andar4 = view.findViewById(R.id.imageViewAndar4);
         andar5 = view.findViewById(R.id.imageViewAndar5);
+
     }
 
 }
