+++ app/src/main/java/br/com/digitalhouse/harvardartmuseums/model/favorites/Favorites.java
@@ -1,5 +1,6 @@
 package br.com.digitalhouse.harvardartmuseums.model.favorites;
 
+import android.content.Context;
 import android.os.Parcel;
 import android.os.Parcelable;
 
@@ -8,6 +9,8 @@ import androidx.room.PrimaryKey;
 
 import com.google.gson.annotations.Expose;
 
+import br.com.digitalhouse.harvardartmuseums.data.database.Database;
+import br.com.digitalhouse.harvardartmuseums.data.database.dao.FavoritesDAO;
 import br.com.digitalhouse.harvardartmuseums.model.object.Object;
 
 @Entity(tableName = "tab_favorites")
@@ -99,4 +102,5 @@ public class Favorites implements Parcelable {
     public void setObjectGallery(Object objectGallery) {
         this.objectGallery = objectGallery;
     }
+
 }
