+++ app/src/main/res/layout/activity_base.xml
@@ -4,7 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.base.BaseActivity">
+    tools:context=".view.base.BaseActivity"
+    android:background="@color/colorWhite">
 
     <include
         android:id="@+id/toolbar"
