+++ app/src/main/res/layout/activity_setting.xml
@@ -1,17 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="horizontal"
-    tools:context=".view.settings.SettingsActivity">
+    tools:context=".view.settings.SettingsActivity"
+    android:background="@color/colorWhite">
 
     <include
         android:id="@+id/toolbar"
         layout="@layout/toolbar" />
-
 
     <ImageView
         android:id="@+id/imageView1"
@@ -128,7 +127,7 @@
         app:layout_constraintStart_toEndOf="@+id/imageView"
         app:layout_constraintTop_toTopOf="@+id/imageView" />
 
-    <EditText
+    <TextView
         android:id="@+id/device"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
@@ -146,5 +145,6 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/toolbar" />
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
