+++ app/src/main/res/layout/fragment_info.xml
@@ -3,6 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@color/colorWhite"
     tools:context=".fragments.information.InformationActivity">
 
     <LinearLayout
@@ -10,22 +11,13 @@
         android:layout_height="match_parent"
         android:orientation="vertical">
 
-        <TextView
+        <LinearLayout
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/toolbar"
-            android:layout_marginTop="20dp"
-            android:gravity="center"
-            android:text="Each click is an opportunity to interact with us!"
-            android:textColor="#000000" />
+            android:layout_height="wrap_content">
 
-        <ImageView
-            android:layout_width="100dp"
-            android:layout_height="100dp"
-            android:layout_marginStart="16dp"
-            android:layout_marginLeft="16dp"
-            android:layout_marginTop="6dp"
-            android:src="@drawable/guia" />
+            <include layout="@layout/content_top_message"></include>
+
+        </LinearLayout>
 
         <TextView
             android:layout_width="match_parent"
