+++ app/src/main/res/layout/recyclerview_exhibition_item.xml
@@ -17,7 +17,7 @@
             android:layout_marginStart="8dp"
             android:layout_marginLeft="8dp"
             android:layout_marginTop="8dp"
-            android:background="@color/colorPrimary"
+            android:src="@drawable/image_logo_center"
             android:scaleType="centerCrop"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
